refactore: canged the class to function
